Fix(quantum cell) destroy proc was called multiple times, likely due to a qdel loop in the Destroy logic

+++ code/modules/power/cell.dm
@@ -50,7 +50,9 @@
 	return amount_used
 
 /obj/item/weapon/cell/quantum/Destroy()
-	qdel(partner)
+	if(!QDELETED(partner))
+		qdel(partner)
+	partner = null
 	return ..()
 
 /obj/item/weapon/cell/New()
